refactor(data): 调整数据存储类型位置
- 将 CellItemStorage 类从 storage 包移动到 data 包，并重命名为 CellItemStorageData
- 更新 CellItemSavedData 类中的相关引用和方法
- 修改 OreStorageCellInventory 类中的相关引用和方法
- 优化数据存储和读取逻辑
